Class-domain relations: Move model parser to a separate file
